requirements.txt
asgiref==3.5.2
Django==4.1.2
django-discover-runner==1.0
django-role-permissions==3.1.1
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.3
